[REFACTOR] refactor images API
[why]
For testability and maintainability

[what]
image route
image controller

[etc]
